refactor(#356): carve spawn_node into node_spawn with injectable SpawnDeps + unit tests
Move spawn_node + SpawnContext + SpawnOutcome (+ private find_collection_context
and fail_spawn_before_start) out of the 22k-line lib.rs god-file into a new
node_spawn module, next to its callers (mirrors worktree_ops / run_advance).

spawn_node now takes a narrow SpawnDeps<'a> — a struct-of-borrows over exactly
the six side-effects it touches (db, event sink, admission lock, panic flag,
port, tmux override) — instead of the whole 40-field &AppState. SpawnDeps is a
bare-Mutex borrow (not a trait), so the admission guard keeps AppState's
lifetime. Built once via SpawnDeps::from_state at each of the 9 call sites
(7 in lib.rs, 2 in run_advance.rs).

Behaviour-preserving: the body is a byte-for-byte move plus mechanical field
swaps (state.X -> deps.X, append_event -> append_event_with, reload_run_state ->
reload_run_state_with); no statement crosses the ordering fences (admission
lock -> cap check -> create_sub_worktree -> catch_unwind span -> drop guard ->
tmux spawn). panic_payload_message stays in lib.rs (shared with
detach_terminal_tail) and is now pub(crate).

Adds 3 direct-drive unit tests that were impossible before the seam — driving
spawn_node through SpawnDeps::from_state without a live daemon:
  * throttle into waiting at the session cap (INV-3 / #213),
  * panic in the spawn window reaps the orphan sub-worktree and fails the run
    loud as RunFailed, not NodeFailed (INV-4/5/6 / #279),
  * transition guard refuses a concurrent iteration before any side effect
    (INV-1 / #212).

Leaf primitive (ADR-0009): spawn_node appends events + touches tmux/worktree
and never re-enters the scheduler — non-reentrancy preserved. No ADR, no
CONTEXT.md change (SpawnDeps/SpawnOutcome are code-level; DI is already blessed
by ADR-0019).
